+++ engine/flink/components-api/src/main/scala/pl/touk/nussknacker/engine/flink/api/NkGlobalParameters.scala
@@ -52,7 +52,7 @@ object NamespaceMetricsTags {
       NamespaceMetricsTags(
         Map(
           originalNameTag -> scenarioName,
-          namespaceTag    -> namespace
+          namespaceTag    -> namespace.value
         )
       )
     }